refactor list of certifications into its own component
